stable android tests (up to 12970) and some changes for ios
