+++ android/locators/dashboard_locators.py
@@ -24,14 +24,14 @@ DASH_WESHARES_CHART = "com.socialsuperstore:id/chart"
 DASH_TRANS_EMPTY_PURCHASES_STUB = "com.socialsuperstore:id/purchasesEmpty"
 DASH_TRANS_EMPTY_INFLUENCE_SALES_STUB = "com.socialsuperstore:id/influencedEmpty"
 DASH_TRANS_EMPTY_REFERRALS_STUB = "com.socialsuperstore:id/referralsEmpty"
-DASH_TRANS_ALERT_CONTAINER = "com.socialsuperstore:id/alertContainer"
+DASH_TRANS_EMPTY_SHARE_BTN = "com.socialsuperstore:id/shareBtn"
 DASH_TRANS_SHOW_MORE_PURCHASES = "com.socialsuperstore:id/youPurchasesLoadMore"
 DASH_TRANS_SHOW_MORE_PRODUCT_TITLE = "com.socialsuperstore:id/dashboardTransactionTitle"
 DASH_TRANS_SHOW_MORE_PRICE = "com.socialsuperstore:id/dashboardTransactionPrice"
 DASH_TRANS_SHOW_MORE_SALES = "com.socialsuperstore:id/influencedSalesLoadMore"
 DASH_TRANS_SHOW_MORE_REFERRALS = "com.socialsuperstore:id/friendsReferredLoadMore"
 DASH_TRANS_SHOW_MORE_REFERRALS_HEADER = "com.socialsuperstore:id/referralsHeader"
-DASH_TRANS_SHOW_MORE_REFERRALS_NICK_INPUT = "com.socialsuperstore:id/nickInput"		
+DASH_TRANS_SHOW_MORE_REFERRALS_NICK_INPUT = "//*[contains(@resource-id, 'linkInput')]//android.widget.FrameLayout/android.widget.EditText" # "//*[contains(@resource-id, 'linkInput')]//android.widget.EditText"  
 
 # WENEWS TAB
 DASH_WENEWS_READ_MORE_BTN = "com.socialsuperstore:id/newsReadMore"
