+++ android/locators/post_locators.py
@@ -27,12 +27,11 @@ POST_FLAG_CONTENT_SUCCESS_WIN_TITLE = "com.socialsuperstore:id/title"
 POST_FLAG_CONTENT_SUCCESS_WIN_DONE_BTN = "com.socialsuperstore:id/doneButton"
 POST_SUB_MENU_ACTION_ITEMS_ID = "com.socialsuperstore:id/actionText"
 POST_MEDIA_LAYOUT = "com.socialsuperstore:id/postProductsMediaLayout"
-READ_ALL_PRODUCT_LINEAR_LAYOUTS = "//androidx.appcompat.widget.LinearLayoutCompat/androidx.recyclerview.widget.RecyclerView/android.view.ViewGroup[1]/android.widget.HorizontalScrollView/android.widget.LinearLayout/android.widget.LinearLayout"
-READ_SINGLE_PRODUCT_LINEAR_LAYOUTS = "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.view.ViewGroup/androidx.appcompat.widget.LinearLayoutCompat/android.widget.FrameLayout/android.view.ViewGroup/android.view.ViewGroup/androidx.recyclerview.widget.RecyclerView/android.view.ViewGroup/android.widget.HorizontalScrollView/android.widget.LinearLayout/android.widget.LinearLayout"
+READ_ALL_PRODUCT_LINEAR_LAYOUTS = "//*[contains(@resource-id, 'postProductsPagerTabs')]//android.widget.LinearLayout/android.widget.LinearLayout"
+READ_SINGLE_PRODUCT_LINEAR_LAYOUTS = "//*[contains(@resource-id, 'postProductsPagerTabs')]//android.widget.LinearLayout/android.widget.LinearLayout"  #'(//*[contains(@resource-id, "postProductsPagerTabs")]//android.widget.LinearLayout/android.widget.LinearLayout)[1]'
 PRODUCT_EDIT_FIRST_CHECKBOX = "//androidx.recyclerview.widget.RecyclerView/android.widget.FrameLayout[1]/android.view.ViewGroup/android.widget.CheckBox"
 POST_USERNAME = "com.socialsuperstore:id/postUserName"
 POST_TIME_AGO_TEXT = "com.socialsuperstore:id/timeAgoTextView"
-
 QUESTION_TEXT_STEP_ONE = "com.socialsuperstore.feature_post_editor:id/questionTextInputEditText"
 QUESTION_TEXT_MEDIA_TAB = "com.socialsuperstore.feature_post_editor:id/questionText"
 QUESTION_BREAD_CRUMBS = "com.socialsuperstore:id/currentStep"
