+++ android/pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 markers =
-    regression: mark a test as regression.
-    sanity: mark a test as sanity.
+    uat: mark to test on uat.
+    int: mark to test on int.
 
-addopts = --driver BrowserStack --capability browserName Appium -v -s --html=report.html --self-contained-html --disable-pytest-warnings
+addopts = --driver Appium --variables caps.json -v -s --html=report.html --self-contained-html --disable-pytest-warnings
