+++ android/ui_page_objects/dashboard_page_object.py
@@ -56,7 +56,7 @@ class DashboardPage:
 		assert el_id(driver, DASH_TRANS_EMPTY_PURCHASES_STUB).is_displayed()
 		assert el_id(driver, DASH_TRANS_EMPTY_INFLUENCE_SALES_STUB).is_displayed()
 		assert el_id(driver, DASH_TRANS_EMPTY_REFERRALS_STUB).is_displayed()
-		assert el_id(driver, DASH_TRANS_ALERT_CONTAINER).is_displayed()
+		assert el_id(driver, DASH_TRANS_EMPTY_SHARE_BTN).is_displayed()
 
 
 	def existing_acc_check(self, driver):
@@ -87,10 +87,10 @@ class DashboardPage:
 		last_updated = el_id(driver, DASH_WESHARES_LAST_UPDATED).text.split(", ")[1]
 		assert last_updated == CURRENT_DATE
 		
-		share_price = float(el_id(driver, DASH_WESHARES_SHARE_PRICE).text.split("£")[1].replace(",", "."))
+		share_price = round_two_dots(el_id(driver, DASH_WESHARES_SHARE_PRICE).text) #float(el_id(driver, DASH_WESHARES_SHARE_PRICE).text.split("£")[1].replace(",", "."))
 		assert share_price > 0
 
-		your_weshare_top = float(el_id(driver, DASH_WESHARES_YOUR_WESHARE_TOP).text.split("£")[1].replace(",", "."))
+		your_weshare_top = round_two_dots(el_id(driver, DASH_WESHARES_YOUR_WESHARE_TOP).text) #float(el_id(driver, DASH_WESHARES_YOUR_WESHARE_TOP).text.split("£")[1].replace(",", "."))
 		assert your_weshare_top > 100
 
 		your_weshare_bottom = float(el_id(driver, DASH_WESHARES_YOUR_WESHARE_BOTTOM).text)
@@ -108,12 +108,13 @@ class DashboardPage:
 
 		# check purchases
 		click_on_show_more_purchases = id_click(driver, DASH_TRANS_SHOW_MORE_PURCHASES)
-		id_until_gone_short(driver, PRE_LOADER)
+		time.sleep(1)
+		#id_until_gone_short(driver, PRE_LOADER)
 
 		error = False
 
 		try:
-			read_possible_error = get_toast_msg(driver)
+			read_possible_error = get_toast_msg_short(driver)
 			error = True 
 		except:
 			pass 
@@ -125,7 +126,7 @@ class DashboardPage:
 			pass
 
 		purchased_product_title = len(el_id(driver, DASH_TRANS_SHOW_MORE_PRODUCT_TITLE).text)
-		purchased_product_price = float(el_id(driver, DASH_TRANS_SHOW_MORE_PRICE).text.split("£")[1].replace(",", "."))
+		purchased_product_price = round_two_dots(el_id(driver, DASH_TRANS_SHOW_MORE_PRICE).text) #float(el_id(driver, DASH_TRANS_SHOW_MORE_PRICE).text.split("£")[1].replace(",", "."))
 
 		assert purchased_product_title > 3
 		assert purchased_product_price > 0
@@ -135,7 +136,7 @@ class DashboardPage:
 		click_on_show_more_sales = id_click(driver, DASH_TRANS_SHOW_MORE_SALES)
 		id_until_gone_short(driver, PRE_LOADER)
 		influenced_sales_product_title = len(el_id(driver, DASH_TRANS_SHOW_MORE_PRODUCT_TITLE).text)
-		influenced_sales_price = float(el_id(driver, DASH_TRANS_SHOW_MORE_PRICE).text.split("£")[1].replace(",", "."))
+		influenced_sales_price = round_two_dots(el_id(driver, DASH_TRANS_SHOW_MORE_PRICE).text) #float(el_id(driver, DASH_TRANS_SHOW_MORE_PRICE).text.split("£")[1].replace(",", "."))
 
 		assert influenced_sales_product_title > 3
 		assert influenced_sales_price > 0
@@ -145,7 +146,7 @@ class DashboardPage:
 		click_on_show_more_referrals = id_click(driver, DASH_TRANS_SHOW_MORE_REFERRALS)
 		id_until_gone_short(driver, PRE_LOADER)
 		read_referral_header_title = el_id(driver, DASH_TRANS_SHOW_MORE_REFERRALS_HEADER).text
-		read_link_in_share_input = el_id(driver, DASH_TRANS_SHOW_MORE_REFERRALS_NICK_INPUT).text.split("@")[1]
+		read_link_in_share_input = el_xpath(driver, DASH_TRANS_SHOW_MORE_REFERRALS_NICK_INPUT).text.split("@")[1]
 		read_my_invited_referral_name = len(el_id(driver, DASH_TRANS_SHOW_MORE_PRODUCT_TITLE).text)
 
 		assert read_referral_header_title == "Invite your friends"
