+++ android/ui_page_objects/product_detail_page_object.py
@@ -121,7 +121,8 @@ class ProductDetailPage:
 	def open_product_website(self, driver):
 		# going to detail product page
 		click_on_home_footer_btn = acc_id_click(driver, FOOTER_ITEM_HOME)
-		scroll_on_feed_page(driver)
+		#scroll_on_feed_page(driver) # not sure
+		scroll_on_feed_page_more(driver)
 		read_product_title = el_id(driver, FEED_PRODUCT_TITLE).text
 		product_title_click = id_click(driver, FEED_PRODUCT_TITLE)
 		read_product_name_on_detail_page = el_id(driver, PRODUCT_NAME_TITLE).text
@@ -140,7 +141,8 @@ class ProductDetailPage:
 	def add_product_to_post(self, driver):
 		# going to detail product page
 		click_on_home_footer_btn = acc_id_click(driver, FOOTER_ITEM_HOME)
-		scroll_on_feed_page(driver)
+		#scroll_on_feed_page(driver) # not sure
+		scroll_on_feed_page_more(driver)
 		click_on_first_product_in_feed = id_click(driver, FEED_PRODUCT_TITLE)
 
 		# product detail page steps > add to post
@@ -157,7 +159,8 @@ class ProductDetailPage:
 	def add_product_to_question(self, driver):
 		# going to detail product page
 		click_on_home_footer_btn = acc_id_click(driver, FOOTER_ITEM_HOME)
-		scroll_on_feed_page(driver)
+		#scroll_on_feed_page(driver)
+		scroll_on_feed_page_more(driver) # fine
 		click_on_first_product_in_feed = id_click(driver, FEED_PRODUCT_TITLE)
 
 		# product detail page steps > add to question
